+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.8.3 2023-09
 
+* [Plugin API add openWindow and command.globalCallback](https://github.com/siyuan-note/siyuan/issues/9032)
+
 ## 0.8.2 2023-08-29
 
 * [Improve IModel type definitions](https://github.com/siyuan-note/petal/pull/14)